requirements.txt
faiss-cpu>=1.7.4
numpy>=1.24.0
openai>=1.0.0
tqdm>=4.65.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
qdrant-client>=1.7.0
lxml>=5.0.0
pytest>=7.4.0
httpx>=0.25.0
